connect invoice CLIENT to SERVER

+++ services/Helper.js
@@ -1,19 +1,56 @@
 import redirect from './redirect';
 import { removeCookie } from './Cookies';
+import fetch from "isomorphic-unfetch";
+import { toast } from 'react-toastify';
 
 
 
-export const apiBaseUrl = 'http://localhost:8000/api';
+export const apiBaseUrl = 'https://e-invoice-app.herokuapp.com/api';
 
 
 
 // Sign Out
-export const signOut = (ctx = {}) => {
+export const signOut = async (ctx = {}) => {
   if (process.browser) {
-    removeCookie('token');
-    redirect("/login", ctx);
+
+    // make a request to the server
+    try{
+      const response = await fetch(`${apiBaseUrl}/logout`, {
+          method: 'POST',
+          headers: { 
+            Accept: "application/json",
+            "Content-Type": "application/json",
+            Authorization: "Bearer " + localStorage.getItem('authToken'),
+          
+          },
+          
+      })
+      if (response.ok) {
+          const data = await response.json();
+          const message = data['message'];
+          toast.success(message, { autoClose: 7000 });
+          removeCookie('token');
+          localStorage.clear();
+          redirect("/login", ctx);
+          
+      } else {
+          toast.warning('Could Not logout...Please try again', { autoClose: 5000 });
+          // https://github.com/developit/unfetch#caveats
+          let error = new Error(response.statusText)
+          error.response = response
+          return Promise.reject(error)
+      }
+    } catch (error) {
+        toast.error('Hmmm...Something Went Wrong', { autoClose: 5000 });
+        console.error(
+            'You have an error in your code or there are Network issues.',
+            error
+        )
+        throw new Error(error);
+    }
+
+   
   }
-  localStorage.clear();
 };
 
 // get token
@@ -31,4 +68,18 @@ export const getUserId = ()  => {
  }
 }
 
-// const token = localStorage.getItem('authToken');
+export const showApiRequestError = (message, res) => {
+  toast.warning(message, { autoClose: 5000 });
+  console.log("Error fetching data");
+  let error = new Error(res.statusText);
+  error.response = res;
+  return Promise.reject(error);
+}
+
+// Make Get Request
+// export const getApiRequest = (route) => {
+//   if (process.browser) {
+
+//   }
+// }
+// export const token = localStorage.getItem('authToken');
